#2005: extract install flow + capacity limits into session/install.rs
Pure code-motion split of the conntrack session-creation path. Move the
install/upsert family and the #1861 capacity preflight off SessionTable's
monolithic impl in session/mod.rs into a focused session/install.rs
submodule: can_admit / note_admission_refused / admission_refused /
note_install_partial / install_partial / create_drops /
set_max_sessions_for_test (capacity), install_with_protocol /
install_with_protocol_with_origin / upsert_synced /
upsert_synced_with_origin (installs), and emit_open_delta_with_origin /
emit_close_delta_with_origin / delete / demote_owner_rg (delta-emit +
delete + owner-RG demotion). Bodies are byte-for-byte identical; only the
file boundary changed.

No visibility was widened: every moved method keeps its original
`pub`/`pub(crate)` visibility (callable crate-wide off the pub(crate)
SessionTable type). The private helpers/structs these methods touch —
remove_entry, next_epoch, index_forward_nat_key, push_delta,
entry_by_key/entry_by_key_mut, session_timeout_ns, and the
SessionRecord/SessionEntry structs — stay in mod.rs and are visible to
this descendant; push_to_wheel (pub(in crate::session) in expire.rs) and
owner_rg_session_keys (pub in lookup.rs) are reached via their existing
visibility.

These installs build fresh records (remove_entry + slab insert +
index_forward_nat_key + push_to_wheel), NOT the in-place-refresh path.
The #1752/#1855 in-place-refresh contract (update_session,
refresh_for_ha_transition, the secondary-index re-assert helpers, and
the #964 eager-cleanup remove_entry) stays untouched in mod.rs.

Validation: cargo build --release clean; cargo test --release session::
78 session tests pass. A function-body diff against the pre-split mod.rs
confirms every moved body is byte-identical and the only retained-code
change to mod.rs is the trimmed `use wheel::SessionWheel` + the three
`mod` declarations.
