+++ userspace-dp/src/session/install.rs
@@ -1,0 +1,309 @@
+// Install flow + capacity limits (#2005 pure code-motion split of
+// session/mod.rs). The session-creation paths and their pre-flight
+// admission live here: the #1861 capacity preflight/counter accessors
+// (can_admit, note_admission_refused, admission_refused,
+// note_install_partial, install_partial, create_drops,
+// set_max_sessions_for_test), the new-flow installs (install_with_protocol,
+// install_with_protocol_with_origin, upsert_synced,
+// upsert_synced_with_origin), and the delta-emit / delete / owner-RG
+// demotion helpers (emit_open_delta_with_origin,
+// emit_close_delta_with_origin, delete, demote_owner_rg). Bodies are
+// byte-for-byte identical; only the file boundary changed.
+//
+// No visibility was widened: every moved method keeps its original `pub`
+// / `pub(crate)` visibility (callable crate-wide off the pub(crate)
+// SessionTable type), and the private helpers/structs these methods
+// touch — remove_entry, next_epoch, index_forward_nat_key, push_delta,
+// entry_by_key_mut, session_timeout_ns, SessionRecord/SessionEntry — live
+// in the parent mod.rs and are visible to this descendant; push_to_wheel
+// (in expire.rs) and owner_rg_session_keys (in lookup.rs) are reached via
+// their existing pub/pub(in crate::session) visibility.
+//
+// The #1752/#1855 in-place-refresh contract (update_session,
+// refresh_for_ha_transition, and the secondary-index re-assert helpers)
+// stays in mod.rs — these installs build fresh records (remove_entry +
+// insert), they are not the in-place-refresh path.
+
+use super::*;
+
+impl SessionTable {
+    /// #1861 §5.1: pre-flight admission for an install group of `needed`
+    /// new entries within ONE descriptor iteration. The table is
+    /// per-worker, single-threaded, and `&mut`-exclusive during packet
+    /// processing (GC and worker commands run between poll phases), so a
+    /// passing preflight guarantees the subsequent `needed` installs
+    /// cannot fail the `max_sessions` cap — the only install failure
+    /// mode. Conservative on purpose: charges a full slot per entry even
+    /// when the key already exists (a replacement would not grow the
+    /// table). That matches `install_with_protocol_with_origin`'s own cap
+    /// check, which also refuses replacements at cap; crediting
+    /// replacements here would let the preflight pass while the install
+    /// itself fails, violating the post-preflight infallibility contract.
+    #[inline]
+    pub fn can_admit(&self, needed: usize) -> bool {
+        self.len().saturating_add(needed) <= self.max_sessions
+    }
+
+    /// #1861 §5.1: counted preflight refusal (one per refused flow).
+    pub fn note_admission_refused(&mut self) {
+        self.admission_refused = self.admission_refused.saturating_add(1);
+    }
+
+    /// #1861: cumulative pair-admission preflight refusals.
+    pub fn admission_refused(&self) -> u64 {
+        self.admission_refused
+    }
+
+    /// #1861 §5.2: counted impossible-by-construction release residual —
+    /// an install failed after a passing preflight. Call sites pair this
+    /// with `debug_assert!(false, ...)` per the #1855 contract.
+    pub fn note_install_partial(&mut self) {
+        self.install_partial = self.install_partial.saturating_add(1);
+    }
+
+    /// #1861: cumulative post-preflight partial-install residuals.
+    pub fn install_partial(&self) -> u64 {
+        self.install_partial
+    }
+
+    /// #1861: cumulative at-cap install refusals from
+    /// `install_with_protocol_with_origin` (previously write-only —
+    /// at-cap drops were operator-invisible).
+    pub fn create_drops(&self) -> u64 {
+        self.create_drops
+    }
+
+    /// #1861 tests: shrink the cap so at-cap interleavings can be rigged
+    /// without inserting 131k entries.
+    #[cfg(test)]
+    pub(crate) fn set_max_sessions_for_test(&mut self, n: usize) {
+        self.max_sessions = n;
+    }
+
+    #[cfg_attr(not(test), allow(dead_code))]
+    pub fn install_with_protocol(
+        &mut self,
+        key: SessionKey,
+        decision: SessionDecision,
+        metadata: SessionMetadata,
+        now_ns: u64,
+        protocol: u8,
+        tcp_flags: u8,
+    ) -> bool {
+        self.install_with_protocol_with_origin(
+            key,
+            decision,
+            metadata,
+            SessionOrigin::ForwardFlow,
+            now_ns,
+            protocol,
+            tcp_flags,
+        )
+    }
+
+    // NOTE: this fn stays positional (NOT context-struct) per the
+    // #1357 plan v3.1 rollback criterion. Empirical codegen showed
+    // the struct destructuring prelude added ~30% standalone-copy
+    // instructions (274 -> 357), exceeding the plan's 10% gate. The
+    // 7 non-inlined production callers (forwarding, poll_descriptor,
+    // session_glue, shared_ops) are on the session-install hot path
+    // and cannot absorb that overhead.
+    pub fn install_with_protocol_with_origin(
+        &mut self,
+        key: SessionKey,
+        decision: SessionDecision,
+        metadata: SessionMetadata,
+        origin: SessionOrigin,
+        now_ns: u64,
+        protocol: u8,
+        tcp_flags: u8,
+    ) -> bool {
+        if self.len() >= self.max_sessions {
+            self.create_drops = self.create_drops.saturating_add(1);
+            return false;
+        }
+        // remove_entry's three debug_assert!s catch invariant
+        // violations in tests:
+        //   - stale-handle guard (entries.get returned None for a
+        //     handle in key_to_handle)
+        //   - PRIMARY-KEY GUARD (record.key != lookup key)
+        //   - no_index_points_at (a secondary index still points
+        //     at the freed handle after cleanup)
+        // The first two guards restore the prior key_to_handle
+        // mapping internally before returning None. In release we
+        // proceed to insert the new record — the guard already
+        // restored the prior mapping; the subsequent
+        // self.key_to_handle.insert(...) overwrites it cleanly.
+        let _previous = self.remove_entry(&key);
+        let epoch = self.next_epoch();
+        let record = SessionRecord {
+            key: key.clone(),
+            entry: SessionEntry {
+                decision,
+                metadata: metadata.clone(),
+                origin,
+                install_epoch: epoch,
+                last_seen_ns: now_ns,
+                expires_after_ns: session_timeout_ns(protocol, tcp_flags, &self.timeouts),
+                closing: matches!(protocol, PROTO_TCP) && (tcp_flags & (TCP_FIN | TCP_RST)) != 0,
+                wheel_tick: 0,
+            },
+        };
+        let raw = self.entries.insert(record);
+        let handle: u32 = raw.try_into().expect("slab handle exceeds u32");
+        self.key_to_handle.insert(key.clone(), handle);
+        self.index_forward_nat_key(&key, handle, decision, &metadata);
+        // #965: schedule the new entry for expiration check.
+        self.push_to_wheel(&key, now_ns);
+        if !metadata.is_reverse && !origin.is_peer_synced() && !origin.is_transient_local_seed() {
+            self.push_delta(SessionDelta {
+                kind: SessionDeltaKind::Open,
+                key,
+                decision,
+                metadata,
+                origin,
+                fabric_redirect_sync: false,
+            });
+        }
+        true
+    }
+
+    #[cfg_attr(not(test), allow(dead_code))]
+    pub fn upsert_synced(
+        &mut self,
+        key: SessionKey,
+        decision: SessionDecision,
+        metadata: SessionMetadata,
+        now_ns: u64,
+        protocol: u8,
+        tcp_flags: u8,
+        allow_replace_local: bool,
+    ) -> bool {
+        self.upsert_synced_with_origin(
+            SessionInstall {
+                key,
+                decision,
+                metadata,
+                origin: SessionOrigin::SyncImport,
+                now_ns,
+                protocol,
+                tcp_flags,
+            },
+            allow_replace_local,
+        )
+    }
+
+    #[inline]
+    pub fn upsert_synced_with_origin(
+        &mut self,
+        req: SessionInstall,
+        allow_replace_local: bool,
+    ) -> bool {
+        let SessionInstall {
+            key,
+            decision,
+            metadata,
+            origin,
+            now_ns,
+            protocol,
+            tcp_flags,
+        } = req;
+        // Reject peer data that would clobber a locally-owned session
+        // unless explicitly allowed (e.g. during HA activation).
+        if matches!(self.entry_by_key(&key), Some(existing) if !existing.origin.is_peer_synced())
+            && !allow_replace_local
+        {
+            return false;
+        }
+        // Same guard semantics as install_with_protocol_with_origin:
+        // remove_entry has 3 debug_assert!s (stale-handle,
+        // primary-key, no_index_points_at) that catch invariant
+        // violations in tests. The first two restore the prior
+        // key_to_handle mapping internally before returning None.
+        let _previous = self.remove_entry(&key);
+        let epoch = self.next_epoch();
+        let record = SessionRecord {
+            key: key.clone(),
+            entry: SessionEntry {
+                decision,
+                metadata: metadata.clone(),
+                origin,
+                install_epoch: epoch,
+                last_seen_ns: now_ns,
+                expires_after_ns: session_timeout_ns(protocol, tcp_flags, &self.timeouts),
+                closing: matches!(protocol, PROTO_TCP) && (tcp_flags & (TCP_FIN | TCP_RST)) != 0,
+                wheel_tick: 0,
+            },
+        };
+        let raw = self.entries.insert(record);
+        let handle: u32 = raw.try_into().expect("slab handle exceeds u32");
+        let index_key = key.clone();
+        self.key_to_handle.insert(key, handle);
+        self.index_forward_nat_key(&index_key, handle, decision, &metadata);
+        // #965: schedule the synced entry for expiration check.
+        self.push_to_wheel(&index_key, now_ns);
+        true
+    }
+
+    pub fn emit_open_delta_with_origin(
+        &mut self,
+        key: SessionKey,
+        decision: SessionDecision,
+        metadata: SessionMetadata,
+        origin: SessionOrigin,
+        fabric_redirect_sync: bool,
+    ) {
+        if metadata.is_reverse {
+            return;
+        }
+        self.push_delta(SessionDelta {
+            kind: SessionDeltaKind::Open,
+            key,
+            decision,
+            metadata,
+            origin,
+            fabric_redirect_sync,
+        });
+    }
+
+    pub fn emit_close_delta_with_origin(
+        &mut self,
+        key: SessionKey,
+        decision: SessionDecision,
+        metadata: SessionMetadata,
+        origin: SessionOrigin,
+    ) {
+        if metadata.is_reverse {
+            return;
+        }
+        self.push_delta(SessionDelta {
+            kind: SessionDeltaKind::Close,
+            key,
+            decision,
+            metadata,
+            origin,
+            fabric_redirect_sync: false,
+        });
+    }
+
+    pub fn delete(&mut self, key: &SessionKey) {
+        self.remove_entry(key);
+    }
+
+    pub fn demote_owner_rg(&mut self, owner_rg_id: i32) -> Vec<crate::session::SessionKey> {
+        if owner_rg_id <= 0 {
+            return Vec::new();
+        }
+        let mut demoted_keys = Vec::new();
+        for key in self.owner_rg_session_keys(&[owner_rg_id]) {
+            let Some(entry) = self.entry_by_key_mut(&key) else {
+                continue;
+            };
+            if !entry.origin.is_peer_synced() {
+                entry.origin = SessionOrigin::SyncImport;
+            }
+            demoted_keys.push(key);
+        }
+        demoted_keys
+    }
+}
